Try to sign a new spk

+++ .sandstorm/setup.sh
@@ -1,4 +1,12 @@
 #!/bin/bash
+
+# When you change this file, you must take manual action. Read this doc:
+# - https://docs.sandstorm.io/en/latest/vagrant-spk/customizing/#setupsh
+
+set -euo pipefail
+
+CURL_OPTS="--silent --show-error"
+
 cd /opt/
 
 PACKAGE=meteor-spk-0.1.8
@@ -7,14 +15,17 @@ CACHE_TARGET="/host-dot-sandstorm/caches/${PACKAGE_FILENAME}"
 
 # Fetch meteor-spk tarball if not cached
 if [ ! -f "$CACHE_TARGET" ] ; then
-    curl https://dl.sandstorm.io/${PACKAGE_FILENAME} > "$CACHE_TARGET"
+    echo -n "Downloading ${PACKAGE}..."
+    curl $CURL_OPTS https://dl.sandstorm.io/${PACKAGE_FILENAME} > "$CACHE_TARGET.partial"
+    mv "${CACHE_TARGET}.partial" "${CACHE_TARGET}"
+    echo "...done."
 fi
 
 # Extract to /opt
 tar xf "$CACHE_TARGET"
 
 # Create symlink so we can rely on the path /opt/meteor-spk
-ln -s "${PACKAGE}" meteor-spk
+if [ ! -e meteor-spk ] ; then ln -s "${PACKAGE}" meteor-spk ; fi
 
 # Add bash, and its dependencies, so they get mapped into the image.
 # Bash runs the launcher script.
@@ -33,13 +44,14 @@ METEOR_CACHE_TARGET="/host-dot-sandstorm/caches/${METEOR_TARBALL_FILENAME}"
 
 # Fetch meteor tarball if not cached
 if [ ! -f "$METEOR_CACHE_TARGET" ] ; then
-    curl "$METEOR_TARBALL_URL" > "${METEOR_CACHE_TARGET}.partial"
+    echo -n "Downloading Meteor version ${METEOR_RELEASE}..."
+    curl $CURL_OPTS "$METEOR_TARBALL_URL" > "${METEOR_CACHE_TARGET}.partial"
     mv "${METEOR_CACHE_TARGET}"{.partial,}
+    echo "...done."
 fi
 
 # Extract as unprivileged user, which is the usual meteor setup
 cd /home/vagrant/
 su -c "tar xf '${METEOR_CACHE_TARGET}'" vagrant
 # Link into global PATH
-ln -s /home/vagrant/.meteor/meteor /usr/bin/meteor
-
+if [ ! -e /usr/bin/meteor ] ; then ln -s /home/vagrant/.meteor/meteor /usr/bin/meteor ; fi
